requirements.txt
opencv-python
numpy
pyyaml
matplotlib
